Add bulk classification plugin
Bulk classification is not a PlugInFilter acting on the open picture, but a regular PlugIn that requests a folder with the target images to be opened first. It then hands the ImageProcessors of all the images as a list to the Classification process.

Changed OKActionListener taking an ImageProcessor as Parameter to taking the List of ImageProcessors and added chained constructor taking a single ImageProcessor that wraps it into a List and passes it to the main constructor.